requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.7.1
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
ta==0.10.2
python-binance==1.0.19
python-dotenv==1.0.0
gunicorn==21.2.0